Rebase

+++ ArduCopter/mode_throw.cpp
@@ -277,13 +277,13 @@ bool ModeThrow::throw_detected()
 
 
     bool possible_throw_detected;
-    switch (g2.throw_type) {
-    case ThrowType::Frisbee:
+    switch ((int) g2.throw_type) {
+    case (int) ThrowType::Frisbee :
         // rotation speed and increasing height indicate a possible throw release
         possible_throw_detected = (frisbee_throw && no_throw_action && changing_height);
         break;
-    case ThrowType::Upward:
-    case ThrowType::Drop:
+    case (int) ThrowType::Upward:
+    case (int) ThrowType::Drop:
         // High velocity or free-fall combined with increasing height indicate a possible air-drop or throw release
         possible_throw_detected = ((free_falling || high_speed) && changing_height && no_throw_action);
         break;
